fix spring annotation

+++ app/src/main/java/com/uade/screenspace/entity/Rating.java
@@ -2,15 +2,15 @@ package com.uade.screenspace.entity;
 
 public class Rating {
 
-    private float rating;
+    private double rating;
     private Comment comment;
     private User user;
 
-    public float getRating() {
+    public double getRating() {
         return rating;
     }
 
-    public void setRating(float rating) {
+    public void setRating(double rating) {
         this.rating = rating;
     }
